Seperate my methods from the Class
# This is for replacing the ElectronGoogleOAuth2 with the original library
